day3_crawl_data/requirements.txt
beautifulsoup4==4.12.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
